Flatten components/layout; move menu items into lib/config.ts
- components/layout/{nav,footer,action-bar}/ -> components/{nav,footer,action-bar}/
- defaultNavItems/defaultFooterItems -> top-level navItems/footerItems exports in lib/config.ts (alongside siteConfig)
- delete the two menu-data.ts files
- update docs anatomy + skills + rollout log entry to match
