+++ apps/docs/content/docs/anatomy/footer.mdx
@@ -26,12 +26,12 @@ The default is an empty array, which hides the social links section entirely.
 
 ## Copyright
 
-The copyright line is a static string. To customize it, edit the `<p>` tag in `components/layout/footer/footer.tsx`.
+The copyright line is a static string. To customize it, edit the `<p>` tag in `components/footer/index.tsx`.
 
 ## Key files
 
 | File | Purpose |
 |------|---------|
-| `components/layout/footer/footer.tsx` | Footer composition, copyright, social links slot |
-| `components/layout/footer/social-links.tsx` | Social media icon links |
-| `lib/config.ts` | `siteConfig.socialLinks` array |
+| `components/footer/index.tsx` | Footer composition, copyright, social links slot |
+| `components/footer/social-links.tsx` | Social media icon links |
+| `lib/config.ts` | `siteConfig.socialLinks`, `navItems`, `footerItems` |
