+++ apps/docs/content/docs/anatomy/navigation.mdx
@@ -44,12 +44,12 @@ A fixed bar at the bottom of the viewport. It renders only when it has children 
 
 | File | Purpose |
 |------|---------|
-| `components/layout/nav/index.tsx` | Main navbar composition |
-| `components/layout/nav/mobile-menu.tsx` | Hamburger menu with left sheet (mobile) |
-| `components/layout/nav/quick-links.tsx` | Quick links (hardcoded by default) |
-| `components/layout/nav/cart.tsx` | Cart icon server component |
-| `components/layout/nav/cart-client.tsx` | Cart icon with quantity badge |
-| `components/layout/nav/search-modal.tsx` | Search icon and modal dialog |
-| `components/layout/footer/footer.tsx` | Footer (copyright only by default) |
-| `components/layout/footer/social-links.tsx` | Social media icon links |
-| `components/layout/action-bar/action-bar.tsx` | Bottom action bar (agent button slot) |
+| `components/nav/index.tsx` | Main navbar composition |
+| `components/nav/mobile-menu.tsx` | Hamburger menu with left sheet (mobile) |
+| `components/nav/quick-links.tsx` | Quick links (hardcoded by default) |
+| `components/nav/cart.tsx` | Cart icon server component |
+| `components/nav/cart-client.tsx` | Cart icon with quantity badge |
+| `components/nav/search-modal.tsx` | Search icon and modal dialog |
+| `components/footer/index.tsx` | Footer (copyright only by default) |
+| `components/footer/social-links.tsx` | Social media icon links |
+| `components/action-bar/index.tsx` | Bottom action bar (agent button slot) |
