--- apps/template/components/layout/footer/menu-data.ts
@@ -1,3 +1,0 @@
-import type { MenuItem } from "@/lib/shopify/types/menu";
-
-export const defaultFooterItems: MenuItem[] = [];
