+++ apps/template/lib/config.ts
@@ -1,3 +1,5 @@
+import type { MenuItem } from "@/lib/shopify/types/menu";
+
 function trimTrailingSlash(value: string): string {
   return value.replace(/\/+$/, "");
 }
@@ -23,6 +25,25 @@ export interface SocialLink {
 
 export const siteConfig = {
   name: process.env.NEXT_PUBLIC_SITE_NAME ?? "Vercel Shop",
-  url: trimTrailingSlash(process.env.NEXT_PUBLIC_BASE_URL || defaultUrl),
   socialLinks: [] as SocialLink[],
+  url: trimTrailingSlash(process.env.NEXT_PUBLIC_BASE_URL || defaultUrl),
 } as const;
+
+export const navItems: MenuItem[] = [
+  {
+    id: "default-nav-shop",
+    title: "Shop",
+    url: "/search",
+    type: "HTTP",
+    items: [],
+  },
+  {
+    id: "default-nav-about",
+    title: "About",
+    url: "/about",
+    type: "HTTP",
+    items: [],
+  },
+];
+
+export const footerItems: MenuItem[] = [];
